move: move post repository to ./post
