+++ src/main/java/com/dku/council/domain/mainpage/service/MainPageServiceImpl.java
@@ -8,9 +8,9 @@ import com.dku.council.domain.mainpage.model.dto.response.CarouselImageResponse;
 import com.dku.council.domain.mainpage.model.dto.response.MainPageResponseDto;
 import com.dku.council.domain.mainpage.model.entity.CarouselImage;
 import com.dku.council.domain.mainpage.repository.CarouselImageRepository;
-import com.dku.council.domain.post.repository.ConferenceRepository;
-import com.dku.council.domain.post.repository.NewsRepository;
-import com.dku.council.domain.post.repository.PetitionRepository;
+import com.dku.council.domain.post.repository.post.ConferenceRepository;
+import com.dku.council.domain.post.repository.post.NewsRepository;
+import com.dku.council.domain.post.repository.post.PetitionRepository;
 import com.dku.council.global.error.exception.IllegalTypeException;
 import com.dku.council.infra.nhn.service.FileUploadService;
 import lombok.RequiredArgsConstructor;
